Issue #82 - Put the behavior to display the dashboard into a component
