+++ frontend/src/app/modules/dashboard/components/dashboard-screen/dashboard-screen.component.css
@@ -1,0 +1,8 @@
+.grid {
+    font-size: 20px;
+    color: #fff;
+    font-family: 'Open Sans', "Helvetica Neue", Helvetica, Arial, sans-serif;
+}
+.grid-item.moving {
+    z-index: 10;
+}
