+++ frontend/src/app/modules/dashboard/components/dashboard-screen/dashboard-screen.component.html
@@ -1,0 +1,17 @@
+<div *ngIf="gridOptions" [ngGrid]="gridOptions" (onItemChange)="updateProjectWidgetsPosition($event)">
+  <div class="commonCssAndJs">
+    <div [innerHTML]="getGridCSS(project.cssStyle) | safeHtml"></div>
+    <div [innerHTML]="getWidgetCommonCSS() | safeHtml"></div>
+  </div>
+  <div #projectWidgetsRendered
+       *ngFor="let projectWidget of this.sortProjectWidgets(project.projectWidgets); let projectWidgetsLoaded = last"
+       [ngClass]="'widget ' + projectWidget.widget.technicalName + ' widget-' + projectWidget.id"
+       [ngGridItem]="{'col': projectWidget.widgetPosition.col,
+                    'row': projectWidget.widgetPosition.row,
+                    'sizey': projectWidget.widgetPosition.height,
+                    'sizex': projectWidget.widgetPosition.width,
+                    'payload': projectWidget.id}"
+       [innerHTML]="getHtmlAndCss(projectWidget) | safeHtml">
+
+  </div>
+</div>
