--- frontend/src/app/modules/dashboard/dashboard-detail/dashboard-detail.component.css
@@ -1,8 +1,0 @@
-.grid {
-  font-size: 20px;
-  color: #fff;
-  font-family: 'Open Sans', "Helvetica Neue", Helvetica, Arial, sans-serif;
-}
-.grid-item.moving {
-  z-index: 10;
-}
