+++ frontend/src/app/modules/dashboard/dashboard.module.ts
@@ -26,6 +26,7 @@ import { DashboardEditComponent } from './dashboard-edit/dashboard-edit.componen
 import { DeleteDashboardDialogComponent } from './components/delete-dashboard-dialog/delete-dashboard-dialog.component';
 import { DeleteProjectWidgetDialogComponent } from './components/delete-project-widget-dialog/delete-project-widget-dialog.component';
 import { EditProjectWidgetDialogComponent } from './components/edit-project-widget-dialog/edit-project-widget-dialog.component';
+import { DashboardScreenComponent } from './components/dashboard-screen/dashboard-screen.component';
 
 const dashboardRoutes: Routes = [
   { path: 'dashboard/:id', component: DashboardDetailComponent, canActivate: [AuthGuard] },
@@ -50,7 +51,8 @@ const dashboardRoutes: Routes = [
       DashboardEditComponent,
       DeleteDashboardDialogComponent,
       DeleteProjectWidgetDialogComponent,
-      EditProjectWidgetDialogComponent
+      EditProjectWidgetDialogComponent,
+      DashboardScreenComponent
   ],
   exports: [
       RouterModule
